requirements.txt
numpy
pandas
matplotlib
toml
torchviz
torchinfo
torchsummary
tensorboard
opencv-python
tqdm
scikit-image
scikit-learn
scipy